+++ CHANGELOG.md
@@ -126,6 +126,11 @@
 - Rename `Pim\Bundle\TransformBundle\Builder\FieldBuilder` to `Pim\Component\Connector\ArrayConverter\Flat\ProductAttributeFieldExtractor`
 - Method `createAttribute` of Pim/Bundle/CatalogBundle/Manager/AttributeManager.php is now deprecated use `AttributeFactory::createAttribute` instead
 - Constructor of `Pim\Bundle\EnrichBundle\Controller\ChannelController` now takes a BulkSaver as last argument (to save locales)
+- Constructor of `Pim\Bundle\CatalogBundle\Manager\AttributeManager` has been changed
+- Constructor of `Pim\Bundle\CatalogBundle\Manager\AttributeGroupManager` has been changed
+- Constructor of `Pim\Bundle\CommentBundle\Manager\CommentManager` has been changed
+- Constructor of `Pim\Bundle\DatagridBundle\Manager\DatagridViewManager` has been changed
+- Constructor of `Pim\Bundle\EnrichBundle\Manager\SequentialEditManager` has been changed
 
 # 1.3.x